moved method getPlayersInLobby from GameController to LobbyController
